backend/requirements.txt
pika
flask
flask-cors
psycopg2-binary
redis
